Description:Codex Code Alarm Handling
Match-id-cd03f6b29a925f5ea3c4625d9312e9a41846209a

+++ services/source/inputinject/include/virtual_device.h
@@ -45,7 +45,7 @@ public:
     std::string GetNetWorkId();
 
     int32_t GetDeviceFd();
-    int32_t GetDeviceType();
+    uint16_t GetDeviceType();
 
 protected:
     VirtualDevice();
